dashboard

+++ application/controllers/Input.php
@@ -10,12 +10,11 @@ class Input extends CI_Controller {
 	}
 	public function index()
 	{
-		// $q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	redirect('login','refresh');
-		// }
+		$q = $this->session->userdata('status');
+		if($q != "login") {
+			redirect('login','refresh');
+		}
 		
-		// $menu['login'] = $this->ModUser->edit($this->session->userdata('admin_id'));
 		$data['input'] = $this->ModInputItems->selectAll();
 		// $data['stok'] = $this->ModItems->getStok();
 		$this->load->view('template/header');
@@ -24,20 +23,20 @@ class Input extends CI_Controller {
 		$this->load->view('template/footer');
 	}
 	public function modal() {
-		// $q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		$q = $this->session->userdata('status');
+		if($q != "login") {
+			exit();
+		}
 		$data['cek'] = 0;
 		$data['item'] = $this->ModInputItems->listItems();
 		$data['supplier'] = $this->ModInputItems->listSupplier();
 		$this->load->view('modal/input', $data);
 	}
 	public function add() {
-		// $q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		$q = $this->session->userdata('status');
+		if($q != "login") {
+			exit();
+		}
 		$id_item = $this->input->post('id_item');
 		$id_supplier = $this->input->post('id_supplier');
 		$qty = $this->input->post('qty_input');
@@ -98,37 +97,13 @@ class Input extends CI_Controller {
 			$this->ModItems->updateStok($total);
 			echo json_encode(array("status" => TRUE));
 		}
-
-		// else {
-		// 	$stok = $this->ModItems->getStok();
-		// 	$total = $stok + $qty;
-		// 	$this->ModItems->updateStok($total);
-		// 	echo json_encode(array("status" => TRUE));
-		// }
-
-		// if ($id_supplier=="0") {
-		// 	$cek = $this->ModSuppliers->cekSupplier();
-		// 	if ($cek == "TRUE") {
-		// 		$this->ModSuppliers->add();
-		// 		echo json_encode(array("status" => TRUE));
-		// 	} else {
-		// 		$this->session->set_flashdata('cekSupplier', 'Supplier sudah terdaftar!');
-		// 	}
-		// }
-		// else {
-		// 	$stok = $this->ModItems->getStok();
-		// 	$total = $stok + $qty;
-		// 	$this->ModItems->updateStok($total);
-		// 	echo json_encode(array("status" => TRUE));
-		// }
-		
 	}
 
 	public function edit($id) {
-		// $q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		$q = $this->session->userdata('status');
+		if($q != "login") {
+			exit();
+		}
 		$data['cek'] = 1;
 		$data['item'] = $this->ModInputItems->listItems();
 		$data['supplier'] = $this->ModInputItems->listSupplier();
@@ -136,35 +111,37 @@ class Input extends CI_Controller {
 		$this->load->view('modal/input', $data);
 	}
 	public function delete($id) {
-		// $q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		$q = $this->session->userdata('status');
+		if($q != "login") {
+			exit();
+		}
 		$this->ModInputItems->delete($id);
 		echo json_encode(array("status" => TRUE));
 	}
 	public function update() {
-		// $q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		$q = $this->session->userdata('status');
+		if($q != "login") {
+			exit();
+		}
 		$this->ModInputItems->update();
 		echo json_encode(array("status" => TRUE));
 	}
 	public function set_supplier($id) {
 		$q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		if($q != "login") {
+			exit();
+		}
 		$data['supplier'] = $this->ModSuppliers->edit($id);
 		$this->load->view('modal/set-supplier', $data);
 	}
 	public function set_item($id) {
 		$q = $this->session->userdata('status');
-		// if($q != "login") {
-		// 	exit();
-		// }
+		if($q != "login") {
+			exit();
+		}
 		$data['items'] = $this->ModItems->edit($id);
 		$this->load->view('modal/set-item', $data);
 	}
+
+
 }
